Split example model landing data
