+++ frontend/lib/examples/modelLandingFaq.ts
@@ -1,0 +1,96 @@
+import type { AppLocale } from '@/i18n/locales';
+import type { ExampleFaqItem } from '@/lib/examples/modelLandingTypes';
+
+const HUB_FAQ_BY_LOCALE: Record<AppLocale, { title: string; items: ExampleFaqItem[] }> = {
+  en: {
+    title: 'Examples FAQ',
+    items: [
+      {
+        question: 'Can I clone these examples directly?',
+        answer: 'Yes. Open an example and reuse its prompt and settings as a starting point in your workspace.',
+      },
+      {
+        question: 'Do examples cover text-to-video AI, image-to-video AI, and video-to-video AI workflows?',
+        answer: 'Yes. The gallery covers text-to-video AI, image-to-video AI, and selected video-to-video AI workflows when the underlying models support them.',
+      },
+      {
+        question: 'Is pricing shown for each example?',
+        answer: 'Pricing is shown per clip so you can compare cost before running similar renders.',
+      },
+      {
+        question: 'Why can two runs differ with the same prompt?',
+        answer: 'Model behavior can vary by mode, settings, and queue context, even with the same prompt.',
+      },
+      {
+        question: 'Where can I compare model specs and limits?',
+        answer: 'Use the models pages for specs/limits and the engine hub for use-case decision workflows.',
+      },
+      {
+        question: 'How do I pick the right engine for this shot?',
+        answer: 'Start from example quality and cost, then validate with a short test run on your target format.',
+      },
+    ],
+  },
+  fr: {
+    title: 'FAQ Exemples',
+    items: [
+      {
+        question: 'Puis-je cloner ces exemples directement ?',
+        answer: 'Oui. Ouvrez un exemple et réutilisez son prompt et ses réglages dans votre studio.',
+      },
+      {
+        question: 'Les exemples couvrent-ils les flux texte-vers-vidéo IA, image-vers-vidéo IA et vidéo-vers-vidéo IA ?',
+        answer: 'Oui. La galerie couvre le texte-vers-vidéo IA, l’image-vers-vidéo IA et certains flux vidéo-vers-vidéo IA lorsque les modèles sous-jacents les prennent en charge.',
+      },
+      {
+        question: 'Le prix est-il affiché pour chaque exemple ?',
+        answer: 'Oui, un prix par clip est affiché pour comparer le coût avant de relancer.',
+      },
+      {
+        question: 'Pourquoi deux runs diffèrent avec le même prompt ?',
+        answer: 'Le comportement varie selon le mode, les réglages et le contexte de rendu.',
+      },
+      {
+        question: 'Où comparer les caractéristiques et limites des modèles ?',
+        answer: 'Utilisez les pages modèles pour les caractéristiques et limites, et la page comparatif pour choisir selon l’usage.',
+      },
+      {
+        question: 'Comment choisir le bon modèle pour ce plan ?',
+        answer: 'Partir du niveau qualité/coût vu en exemples puis valider via un test court au format cible.',
+      },
+    ],
+  },
+  es: {
+    title: 'FAQ de ejemplos',
+    items: [
+      {
+        question: '¿Puedo clonar estos ejemplos directamente?',
+        answer: 'Sí. Abre un ejemplo y reutiliza su prompt y ajustes en tu workspace.',
+      },
+      {
+        question: '¿Los ejemplos incluyen workflows de text-to-video AI, image-to-video AI y video-to-video AI?',
+        answer: 'Sí. La galería cubre text-to-video AI, image-to-video AI y algunos workflows de video-to-video AI cuando los modelos subyacentes los admiten.',
+      },
+      {
+        question: '¿Se muestra el precio por ejemplo?',
+        answer: 'Sí, el precio por clip se muestra para comparar coste antes de generar.',
+      },
+      {
+        question: '¿Por qué dos ejecuciones cambian con el mismo prompt?',
+        answer: 'El comportamiento varía por modo, ajustes y contexto de cola incluso con el mismo prompt.',
+      },
+      {
+        question: '¿Dónde comparo specs y límites por modelo?',
+        answer: 'Usa las páginas de modelos para ver características y límites, y la página comparativa para decidir según el objetivo.',
+      },
+      {
+        question: '¿Cómo elijo el motor correcto para esta toma?',
+        answer: 'Compara calidad y coste en ejemplos y valida con una prueba corta en tu formato objetivo.',
+      },
+    ],
+  },
+};
+
+export function getHubExamplesFaq(locale: AppLocale): { title: string; items: ExampleFaqItem[] } {
+  return HUB_FAQ_BY_LOCALE[locale] ?? HUB_FAQ_BY_LOCALE.en;
+}
